requirements.txt
requests
aiohttp
aiofiles
numpy
scipy
nb-cli
nonebot-adapter-onebot
opencv-python
pillow
tensorboard
torch
torchvision
